poc/a_communication/poc_1_web_socket/integrated/requirements.txt
# Server
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pyjwt>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
